requirements.txt
api==0.0.7
certifi==2024.2.2
charset-normalizer==3.3.2
env==0.1.0
idna==3.6
nose==1.3.7
peewee==3.17.1
pyTelegramBotAPI==4.15.4
python-dotenv==1.0.1
requests==2.31.0
stateful-object==0.1
telebot==0.0.5
Unidecode==1.3.8
urllib3==2.2.0
